% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gov2k1inSilico.R
\docType{package}
\name{Gov2k1inSilico}
\alias{Gov2k1inSilico}
\title{Gov2k1inSilico}
\description{
Develop intuition for fundamental statistical concepts without
having to know any programming languages.
}
\details{
\code{Gov2k1inSilico} is software that presents some of the most important
concepts in statistics (for social science). The app presents the idea of
probability distributions, describes the process of model-based inference,
and demonstrates simulation to derive quantities of interest.
The software makes reference to the freely available lecture videos made by
Gary King for \href{https://projects.iq.harvard.edu/gov2001}{GOV 2001}, Harvard's introductory graduate methodology class
for social scientists.

See the \href{https://projects.iq.harvard.edu/2k1-in-silico/home}{project website}
for more information about the software.
}
