% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGov2k1.R
\name{runGov2k1}
\alias{runGov2k1}
\title{Run the Shiny Application}
\usage{
runGov2k1()
}
\arguments{
\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
